docs: fill empty XML documentation comments
Add meaningful XML doc summaries, param, returns, and typeparam
descriptions across IAgency, GptService, ImageGenerationRequest,
and ImageGenerationModerationException.
