+++ agency/IAgency.cs
@@ -1,5 +1,8 @@
 ﻿namespace ShareInvest.Agency;
 
+/// <summary>
+/// Core contract for AI agency services in the PageMint ecosystem, including batch image generation and title generation.
+/// </summary>
 public interface IAgency
 {
